requirements.txt
nibabel~=5.1.0
numpy~=1.26.1
tensorflow~=2.14.0
monai~=1.3.0
Pillow~=9.3.0
torch~=2.1.0+cu121
SimpleITK~=2.3.0
MedPy~=0.4.0
torchvision~=0.16.0+cu121
scipy~=1.11.3
tqdm~=4.66.1
h5py~=3.10.0
PyYAML~=6.0.1
matplotlib~=3.8.0
scikit-image~=0.22.0